add superadmin gate and add gate to adminusercontroller

+++ app/Http/Controllers/Admin/AdminUserController.php
@@ -2,102 +2,67 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 
 class AdminUserController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
-        // except the current user
+        Gate::authorize('admin')||Gate::authorize('superadmin');
         $users = User::orderBy('name')
-        ->where('id', '!=', '1')
-        ->paginate(20);
+            ->where('id', '!=', '1')
+            ->paginate(20);
         return view('admin.users.index', compact('users'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
+        Gate::authorize('admin')||Gate::authorize('superadmin');
         return view('admin.users.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         //
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Http\Response
-     */
     public function show(User $user)
     {
         //
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Http\Response
-     */
     public function edit(User $user)
     {
         //
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, User $user)
     {
         //
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(User $user)
     {
+        Gate::authorize('admin')||Gate::authorize('superadmin');
         $user->delete();
         return back();
     }
 
     public function makeAdmin(User $user)
     {
+        Gate::authorize('superadmin');
         $user->timestamps = false;
         $user->is_admin = 1;
         $user->save();
         return back();
     }
+
     public function removeAdmin(User $user)
     {
+        Gate::authorize('superadmin');
         $user->timestamps = false;
         $user->is_admin = 0;
         $user->save();
@@ -106,14 +71,16 @@ class AdminUserController extends Controller
 
     public function verify(User $user)
     {
+        Gate::authorize('admin')||Gate::authorize('superadmin');
         $user->timestamps = false;
         $user->is_verified = 1;
         $user->save();
         return back();
     }
+
     public function unverify(User $user)
     {
-        // disable created at and updated at
+        Gate::authorize('admin')||Gate::authorize('superadmin');
         $user->timestamps = false;
         $user->is_verified = 0;
         $user->save();
